done notes

+++ alpha/components/personal/note/page/FolderArea.tsx
@@ -1,11 +1,15 @@
 "use client";
 
+import Confirm from "@/components/error/Confirm";
 import { NoteType } from "@/models/personal/Note";
 import { NoteFolderType } from "@/models/personal/NoteFolder";
 import deleteNoteFolder from "@/utils/delete/deleteNoteFolders";
+import deleteNote from "@/utils/delete/deleteNotes";
 import { useState } from "react";
 import {
+  AiFillFolderOpen,
   AiOutlineDelete,
+  AiOutlineEdit,
   AiOutlineFolder,
   AiOutlineFolderOpen,
   AiOutlineUnorderedList,
@@ -15,6 +19,7 @@ import { FaFile, FaFolder, FaFolderPlus, FaStickyNote } from "react-icons/fa";
 interface Props {
   folders: NoteFolderType[];
   notes: NoteType[];
+  handleDeleteNote: (input: NoteType) => void;
   showNewFolderForm: () => void;
   handleNewNoteView: () => void;
   handleEditNoteView: (note: NoteType) => void;
@@ -31,6 +36,7 @@ const FolderArea: React.FC<Props> = ({
   handleEditNoteView,
   handleReadNoteView,
   handleDeleteFolder,
+  handleDeleteNote,
   setCurrentFolder,
 }) => {
   const [openedFolder, setOpenedFolder] = useState("");
@@ -47,13 +53,16 @@ const FolderArea: React.FC<Props> = ({
     setSelectedNote(note._id);
   };
 
+  const [folderDeleteConfirm, setFolderDeleteConfirm] = useState("");
+  const [noteDeleteConfirm, setNoteDeleteConfirm] = useState("");
+
   const renderNotes = (folderID: string) => {
     const items = notes.filter((note) => note.folder === folderID);
     return (
       <main className="text-xs pl-3 w-full">
         {items.map((item) => (
           <div
-            className={`flex items-center justify-between py-1 my-1 w-full px-1 ${
+            className={`flex items-center justify-between py-1 my-1 w-full px-2 rounded-md ${
               selectedNote === item._id ? "bg-gray-100" : ""
             }`}
             key={item._id}>
@@ -65,19 +74,59 @@ const FolderArea: React.FC<Props> = ({
               </div>
               <span>{item.title}</span>
             </button>
+            <div className="text-sm">
+              <button
+                className="mx-2 p-1 rounded-full hover:bg-gray-200"
+                onClick={() => {
+                  setNoteDeleteConfirm(item._id);
+                }}>
+                <AiOutlineDelete />
+              </button>
+              <button
+                className=" p-1 rounded-full hover:bg-gray-200"
+                onClick={() => handleEditNoteView(item)}>
+                <AiOutlineEdit />
+              </button>
+            </div>
           </div>
         ))}
       </main>
     );
   };
 
-  const handleDeleteClick = async (folder: NoteFolderType) => {
-    const deletedFolder = await deleteNoteFolder(folder._id);
-    handleDeleteFolder(deletedFolder);
+  const handleFolderDeleteClick = async (id: string) => {
+    const deletedFolder = await deleteNoteFolder(id);
+    deletedFolder && handleDeleteFolder(deletedFolder);
+    setFolderDeleteConfirm("");
+  };
+
+  const handleNoteDeleteClick = async (id: string) => {
+    const deletedNote = await deleteNote(id);
+    deletedNote && handleDeleteNote(deletedNote);
+    setNoteDeleteConfirm("");
   };
 
   return (
     <div className="w-full h-full border rounded-sm py-5 px-3 font-semibold">
+      {folderDeleteConfirm && (
+        <Confirm
+          handleBack={() => setFolderDeleteConfirm("")}
+          handleConfirm={() => {
+            handleFolderDeleteClick(folderDeleteConfirm);
+          }}
+          text="delete this folder"
+        />
+      )}
+      {noteDeleteConfirm && (
+        <Confirm
+          handleBack={() => setNoteDeleteConfirm("")}
+          handleConfirm={() => {
+            handleNoteDeleteClick(noteDeleteConfirm);
+          }}
+          text="delete this note"
+        />
+      )}
+
       <section className="w-full">
         <div className=" px-3 flex justify-between items-center pb-3 ">
           <h1 className="font-semibold text-base pl-2">Your Notes</h1>
@@ -111,19 +160,26 @@ const FolderArea: React.FC<Props> = ({
                   onClick={() => {
                     selectFolder(folder);
                   }}>
-                  <AiOutlineFolder className="mr-2 text-lg" />
+                  {selectedFolder === folder._id ? (
+                    <AiFillFolderOpen className="mr-2 text-lg" />
+                  ) : (
+                    <AiOutlineFolder className="mr-2 text-lg" />
+                  )}
+
                   <span>{folder.name}</span>
                 </button>
                 <div className="flex items-center text-sm gap-2">
                   <button
-                    className="p-1.5 rounded-full hover:bg-gray-100"
-                    onClick={() => handleDeleteClick(folder)}>
+                    className="p-1.5 rounded-full hover:bg-gray-200"
+                    onClick={() => {
+                      setFolderDeleteConfirm(folder._id);
+                    }}>
                     <AiOutlineDelete />
                   </button>
                   <button
-                    className={`p-1.5 rounded-full hover:bg-gray-100 ${
+                    className={`p-1.5 rounded-full hover:bg-gray-200 ${
                       openedFolder === folder._id
-                        ? "bg-gray-300 hover:bg-gray-400"
+                        ? "bg-gray-200 hover:font-bold"
                         : ""
                     }`}
                     onClick={() =>
